Clean up

+++ pkg/tasks/c1api/full_sync.go
@@ -184,13 +184,12 @@ func (c *fullSyncTaskHandler) HandleTask(ctx context.Context) error {
 		return c.helpers.FinishTask(ctx, nil, nil, err)
 	}
 
-	err = uploadDebugLogs(ctx, c.helpers)
+	err = uploadDebugLogs(ctx, c.helpers, c.task.GetDebug())
 	if err != nil {
-		l.Error("uploadDebugLogs returned error", zap.Error(err))
+		l.Error("failed to upload debug Task logs", zap.Error(err))
 		return c.helpers.FinishTask(ctx, nil, nil, err)
 	}
 
-	l.Info("HandleTask() returning no error")
 	return c.helpers.FinishTask(ctx, nil, nil, nil)
 }
 
@@ -214,7 +213,9 @@ func newFullSyncTaskHandler(
 	}
 }
 
-func uploadDebugLogs(ctx context.Context, helper fullSyncHelpers) error {
+// Check if Debug logs should be uploaded to C1 and if so do so,
+// otherwise silently return success.
+func uploadDebugLogs(ctx context.Context, helper fullSyncHelpers, deleteDebugLogs bool) error {
 	ctx, span := tracer.Start(ctx, "uploadDebugLogs")
 	defer span.End()
 
@@ -246,7 +247,6 @@ func uploadDebugLogs(ctx context.Context, helper fullSyncHelpers) error {
 		default:
 			l.Warn("cannot stat debug log file", zap.Error(err))
 		}
-		l.Error("debug path stat failed", zap.Error(err))
 		return nil
 	}
 
@@ -255,14 +255,18 @@ func uploadDebugLogs(ctx context.Context, helper fullSyncHelpers) error {
 		l.Error("failed to open debug log file path", zap.Error(err))
 		return err
 	}
-	defer func() {
-		err := os.Remove(debugPath)
-		if err != nil {
-			l.Error("failed to delete file with debug logs", zap.Error(err), zap.String("file", debugPath))
-		} else {
-			l.Info("deleted debug path")
-		}
-	}()
+	if deleteDebugLogs {
+		// We only delete the debug log when asked to,
+		// as Manager-required log files are not automatically rotated.
+		defer func() {
+			err := os.Remove(debugPath)
+			if err != nil {
+				l.Error("failed to delete file with debug logs", zap.Error(err), zap.String("file", debugPath))
+			} else {
+				l.Info("deleted debug path")
+			}
+		}()
+	}
 	defer debugfile.Close()
 
 	l.Info("uploading debug logs", zap.String("file", debugPath))
@@ -272,6 +276,5 @@ func uploadDebugLogs(ctx context.Context, helper fullSyncHelpers) error {
 		return err
 	}
 
-	l.Info("completed upload of debug logs, no error")
 	return nil
 }
